win32: use wcslen() return type

https://bugzilla.gnome.org/show_bug.cgi?id=762202

+++ glib/gwin32.c
@@ -192,7 +192,7 @@ g_win32_error_message (gint error)
 {
   gchar *retval;
   wchar_t *msg = NULL;
-  int nchars;
+  size_t nchars;
 
   FormatMessageW (FORMAT_MESSAGE_ALLOCATE_BUFFER
 		  |FORMAT_MESSAGE_IGNORE_INSERTS
